+++ NEWS.md
@@ -3,11 +3,14 @@ Red Panda C++ Version 2.15
   - fix: Static class members is not correctly recognized as static.
   - fix: Function with reference type return value is not correctly parsed.
   - enhancement: Add descriptions for x86 registers in the cpu info dialog.
-  - fix: Search dialog shouldn't have "confirm when replace".
+  - fix: Search dialog shouldn't have "prompt when replace".
   - change: Default value for the debugger debugger panel "memory view's columns" is changed from 8 to 16.
   - change: Default value for the debugger debugger panel "memory view's rows" is changed from 8 to 16.
   - enhancement: Display hex value as ascii chars in the debugger panel memory view tab.
   - fix: Word on the last line's end can't be searched.
+  - enhancement: Auto close other search/replace dialogs when start to search/replace.
+  - change: Remove "prompt when replace" in the replace.
+  - fix: Search/replace with regex is not correctly handled.
 
 Red Panda C++ Version 2.14